19: Moving ProjectsList component to folder
