+++ src/components/ProjectsList/style.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const StyledProjectsList = styled.ul`
+  padding: 0;
+`;
